requirements.txt
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
tqdm
argparse
Flask
waitress
sanitize_filename
zipfile36
markdown
beautifulsoup4
langchain